Basic file handle with raw IO ops exposed
The raw IO functions (seek, rewind, tell, read) are not really meant to
be used, not even by Python library devs. They're intended to be a
useful porting and experimentation tool, until the python extension
itself is able to do the job.

It's more or less a mirror of the C fstream API. Some utility functions
for easier exception raising are added as well.