Parse EFLR in Python
While the main work is done in the Python extension, an escape hatch is
provided for records that aren't parsed properly, or for ad-hoc rules.

Nothing is done to manage check sums, or to decrypt records. It's
likely that decryption will have to come at a later point, and that
they're driven from python.