Add internal tag function, with iostreams
The tag function is intended to be the workhorse of the mkindex
currently exposed. It's a cleaner implementation of the previous helpers
for mkindex, with slightly different operation.

With that, the std::FILE interface has been replaced with a primitive
file handle based on iostreams. This is likely to be replaced by memory
maps in the future, but for now this provides an exception-aware and
otherwise as convenient file handle. The main benefit with this
implementation is that it's testable without writing on-disk files.