Skip past preceeding frame-data when constant-size
It turns out the parsing of elements preceeding the asked-force curve is
the main bottleneck when extracting a single curve file-wide, which
indicates that the data type parsing routines is a performance killer,
and maybe need some optimisation. Furthermore, constructing and throwing
away python objects is also very expensive.

To get some real world experience with this, hack a primitive, non-core
backed skipping in place if every preceeding element is of constant
size.