Move C++ extension work to lib/
The python extension will largely be written in C++, and it's getting
evident that more dependencies are needed for it. It's practical to
build it with cmake as a non-installed component, and then just link the
extension itself to that, rather than the mess of a massive C++
implementation in the python/dlisio/ext dir.

This patch moves the files into the lib/ tree, and accounts for this in
the makefiles.

Travis needs installed libraries and headers even from the extension to
work, as the python extension is built in a separate step inside a
docker container.