Move bytes-to-C++ functions to python/ext/core
Move the conversion from bytes to C++ types to the conv namespace in the
python extension, to free up the dl namespace for the soon-to-come
strong typedefs for the type itself.

This simplifies the move where these names are used for the actual
types, and the assemble-from-bytes gets redesigned to output strong
typedef'd types.