Parse set template in pure C++
Add the new function parse_template that parses a set template in C++.
The C++ parsing functionality for now lives in lib/src/parse.cpp.

Tons of support functions are needed for this, including some overlap
with that's already available in the python/dlisio/ext/core.cpp, but
it's expected that parse.cpp will fully replace most of the
functionality in core.cpp

This adds a dependency on boost, as the boost::variant underpins the
design.