Parse (some) explicitly formatted records in C++
Support for explicit records in this patch is partial, but sketches out
the likely direction for explicitly formatted record parsing.

In short, model the rules of the RP66 protocol all the way from byte
intepretation as first-class C++ types. It's structurally similar, but a
better implementation, than what powers the python module today.