Read attrib into temporary in channel LONG-NAME
The long-name attribute is a variant of obname-ascii, which defaults to
unset-type. Boost will always raise bad-get when trying to get a
reference to the unset value, which is not uncommon when populating the
channel object from an attribute set.

By giving into a valid object this problem is circumvented, and it's
fine since the type tag already is checked in the channel::set method.