Store attributes as optional, distinguish absent
The logical record types in appendix A have attributes that are commonly
absent, either by not being in the object template, or by having
explicit absent attributes.

This is a perfect fit for option types. This introduces a clear notion
of present/absent attributes:

1. all attributes are absent (None) until set by the object template or
the object template itself
2. attributes for specific objects can be explicitly marked absent

An added benefit of the optional type is that it makes it very clear
that the object attributes are nullable. A consequence is that the
allow_empty parameter to the set method is replaced by a dedicated
remove method (which makes the appropriate attribute absent).