Use C++ objects with bindings for RP66 records
Quit the strategy of building a nested-dict representation of object
layout in python, and replace with a stricter mirror of the RP66
document with C++ objects. In short, replace:

object = {
    'type': 'CHANNEL',
    'name': name(),
    'objects': {
        object_name: { 'label': 'CHANNELS',
                       'reprc': 2,
                       'value': { ... },
        }, ...
    }
}

with:

struct channel {
    ident  label;
    ushort reprc;
    ...
    vector<value> values;
};

py::class_< channel >( m, "channel" )
    .def_readonly( label, &channel::label)
    ...
;

This representation is stricter, and understanding of record semantics
is easier with stronger types and a more rigid structure.

A strong driver for this implementation is the extensive use of
variants, powered by the now-enabled mpark.variant, a backport of C++17
std::variant.