Add dlis_packf and size friends
The dlis_packf function is intended to be the workhorse that extracts
frame data and other arbitrary sequenced data types that are too messy
to manually do in library and application logic.

RP66 has a pattern of a data manifest (or parsing guide, or structured
formatting) which is later used to interpret byte sequences in other
aspects of the file. To address this, the format string seems like a
good fit, i.e. some routines inspect the meta data, such as channel
objects, and produce and store format strings that will later be used to
parse log data.