Add function to find offsets of logical records
This function exists in various level of detail in the tree already,
finally implement it in the core module.

Walking through a file and indexing the logical records is very useful,
because it enables random access through a lookup table. Since virtually
all implementations targeting dlisio would in one way or another want
this, it's provided by the core module.