Initial I/O device to power the python lib
dl::stream is intended to be a a simple, RP66 record aware I/O device.
It's equipped with an escape hatch (the read function) to support raw
I/O whenever the source file is too broken, but otherwise every
operation works on complete records.

This is based on the various catrecord implementations littered around
the tree, which are long overdue for removal.