Add mandreyel/mio for memory-mapping files
mio is a nice alternative to boost.iostreams, when iostreams is only
used for memory mapping. Like mpark-variant, just vendor the full
amalgamated header into the source tree. mio has cmake support and it's
perfectly possible to git-submodule it, but said cmake support requires
cmake >= 3.8, an upgrade which is currently out of scope.