Error dlis-index-records when a len is too small
2.3.6.4 Minimum Visible Record Length specifies minimum record lengths
for both visible records and logical segments, and without failing on
these being too small (most notably zero), this function would carry on
considering everything a success.

This on some particularly bad files would lead to an infinite loop of
allocations and retries, until memory is exhausted or the process is
otehrwise killed.