Add fmtlib (5.3.0) as a submodule
fmtlib [1] is a nifty library for python-style format strings in C++. It
could both help driving the packing features, and help with a lot of the
string formatting going on in dlisio.

fmtlib is already available in debian and similar, but often at later
versions and will still be a problem on CI services etc. so providing it
in a submodule is a reasonable alternative whenever it's not already
available on a system.

[1] http://fmtlib.net