Translate broken degree symbol, handle utf8 errors
Python and pybind assumes utf8 strings, but some files seem to use a
broken encoding of the degree symbol [1] which causes UTF-8 conversion
to fail.

Instead of always returning ascii/ident/unit as string, first try to
patch this by looking for 0xB0 and replacing it with 0xC2 0xB0. If that
still doesn't make it a valid UTF-8 string - return it as bytes.

[1] https://stackoverflow.com/questions/8732025/why-degree-symbol-differs-from-utf-8-from-unicode