Move stripspaces() to basic_object
variable-lenght string types such as OBNAME, ASCII and IDENT is often
padded with trailing whitespaces. This makes for awkward
string-representations in Python. For convenience, string attributes are
striped for both leading and trailing whitespaces, with Python's
str.strip().

Previously striping was only applied to unknown-object, but as this may
be the case for any object-type it is moved to basic_object