Speed up linking by storing obj-refs in dict
It turns out filtering objects was not fast enough: opening and linking
objects in a file took, on my machine, approx 17 seconds on a 3.5G file.

The main culprit was traversing objects in a cubic manner. Instead of
doing this, cache objects by their fingerprint (really the objref type
from the standard) and look them up in an efficient manner that way. It
takes the time from loading a file down to 1.2s (from ~17s).

This patch really is only the proof-of-concepts that outlines expected
performance. The entire approach design of the RP66 objects should be
reviewed and changed to a better approach.