Use RP66 type names in python Objects
Use the type names from RP66. This gives a tiny bit of predictability
for people that know the standard, but this is not the main draw because
the type-as-string is an implementation detail anyway.

What's more interesting is that they now correspond to what's stored in
objrefs, an object fingerprint. This way, fingerprints can both be built
in python and read as-is from the file.