Move link into the objects, from the pool
Teach the objects to link, rather than patching them from the outside.
Of course, this functionality still relies on a global index of the
available objects, but that conveniently built in the pool anyway (and
is necessary regardless of where the link is implemented).

This accomplishes some more encapsulation, making the the objects more
independent and an organic part of the system. Other than that, it
weakens the coupling between the object pool and the objects by simply
keeping the knowlege on how to link internal, instead of patching it
from the outside. This should improve clarity and testability.