Add operator == and != for most types
It is useful, and sometimes necessary, to check data for equality. The
most compelling case is when an object set has two equal objects in it,
where checking for equality allows deciding to throw, ignore, or warn.

This includes equality operators for the strong typedefs. They already
existed, but were not exposed.