Move fingerprint function to the C core
A translation from components into an object reference fingerprint is
useful in a lot of situations, and a good fit for the C core. Some
experiments demonstrate that it's an obvious candidate for a free
function in the Python module too, which means it's beneficial to move
the authorative fingerprint to the core library.

Note that the does not actually test the fingerprint itself, as it's
not guaranteed to move.