Two-phase init objectpool, store objects in dict
Replace the list storage in the object pool with a dict. Since every
object has a uniquely identifying key, the fingerprint, the dict is a
more natural data structure, and fits nicely into the python paradigm.

This patch is fairly large for that kind of structural change, but the
move to the dict reveals that some sets contain duplicated objects. The
wording in RP66 seems to open up for this, provided that the objects
are identical.

This raises an interesting situation - what if they aren't? For now,
log it as an error and use the last-seen version of the object,
deferring the choice of failure to the caller. If two objects are
equal, log it to info and then continue, effectively ignoring it.

This change broke a lot of tests. Arguably they tested slightly
uninteresting and/or broken behaviour, and they make even less sense
now, because there's no order guarantee on any objects anymore, only
that they're all there. This guarantee was never intended to be there in
the first place, so this change make implementation consistent with the
(lack of) ordering in the attributes.