Build index of FDATA records in load
Pre-compute where the FDATA records. While they seem to always dominate
the number of records in a file, an index allow reducing the number of
records traversed for different frame types, which is quite useful.

This is done on the mmap handle as a part of load currently, as this is
*vastly* more performant than using the file.at infrastructure (which
effectively reads all records). On my machine, this reduces the runtime
of the following script on a 3.5G file from ~20s to ~10s.

import dlisio
with dlisio.load(filename) as f:
    key = dlisio.core.fingerprint('FRAME', 'id', 0, 0)
    curves = f.curves(key)
    print(curves.shape)
    print(curves['MNEMONIC'][0])