Kill obname as name-attribute in Python objects
With the introduction of fingerprint, the benefits of directly using
core.obname as the name-attribute in BasicObject has disappered. At this
point it is more convenient to have seperate attrubutes for id, origin
and copynumber:

        name.id         -> name
        name.origin     -> origin
        name.copynumber -> copynumber

A side-effect of removing the underlying datastructure obname in Python is
that its now possible to manually create objects directly from python
without having a cpp-backing for the object. This is especially usefull
for testing and for manual patching of broken files.