Extend pack_varsize/size to unwidened sizes
These function now have two output parameters, src and dst, to describe
both the size of the on-disk format string, and the possibly-widened
in-memory format string.

This replaces the ad-hoc size descriptor function in the extension.
Tests aren't updated except for in signatures, i.e. it compiles and
passes, but the new (disk-size query) is not tested.