Make object_attributes dict
Replace the object_attribute binding with a dict, and populate it in
C++.

Pybind11 created objects are fairly slow to work with, and attribute
assignment is fairly slow. These objects are accessed a lot, but for
anything but object really only once in most runs.

The count and reprc attributes are not carried over to python, as the
list or value in attr['value'] already contains this information. Units
are omitted as they're not really used right now.

While this make object_attributes slightly more ad-hoc and informaly
specified, this is a huge speed win. On a ~35M test file with about 12M
of explicit records (and most of those vendor-specific), loading the
file is taken from ~3.6s to ~1.8s.