Re-format dtype.names on duplicate mnemonics
When a frame has multiple channels with the same name, enrich the label
by adding :origin:copynumber, so that all names given to numpy.dtype are
unique. This is already the built-in disambiguation mechanism of RP66.

This addition is not done by default because for the cases with only one
mnemonic for the frame, the added formatting is obnoxious and
unnecessary.