Run setup.py with --skip-cmake for travis deploy
The wheels are properly built during all the other stages - deploy is
only meant to upload. However, if setup.py is invoked without
--skip-cmake, it will see that the flags are different and re-run cmake,
which will fail the build.

The failed build is actually not a problem for the uploads (at least not
of now) because the error code is not checked, and the wheel is already
available and uploaded as intended. However, it is subtly wrong, and
if the build should succeed it would be uploaded some weird
non-manylinux wheel, which is not desirable.