Support user-formatted duplicated curve mnemonics
Curve mnemonics can use the full non-blank upper-case ASCII set. While
it is rather uncommon that mnemonics contain full stops, it is desirable
that users can custom-format mnemonics in the case of duplicates to have
a predictably parsable or searchable mnemonic set for that specific
file.

The real change is simply just getting the formatting string from an
instance attribute, rather than a hardcoded string.