Allow unregistered attributes in BasicObject
Previously if object of known type had unexpected attribute, object has
been converted to Unknown. To prevent this stash is introduced.

All the attributes that are unknown to the predefined objects will now
be present in stash.
Moving logic to BasicObject class allows unification of Uknown type with
predefined types.