Unify linkage
Link implementation is changed to stay in basicobject class. Separate
objects now only have to provide knowledge about expected obname
classes.

OBNAME, OBJREF, ATTREF are considered to be links. If object like that
appeared in known attributes, it's moved to refs. Such approach should
simplify dealing with objects that are not always links (like some names
are allowed to be ASCII or OBNAME link to LONG-NAME attributes).

During the linking stage, linkage information is used to populate
attributes with actual references to objects.

If referenced type is not known, links will stay in refs object only and
corresponding field won't be populated.

Handling specific cases (like in Group object where obname type is
defined in other attribute) should not require much refactoring, and
specific logic could be implemented in children classes.

Links in stash are not moved to references as obnames would require
additional knowledge which we do not possess; objref and attrref are not
expected to appear often. Hence this functionality is disregarded for
now.

Co-authored-by: Erlend Hårstad <erha@statoil.com>