Permit reprcode change to the same reprcode
Recently a check has been introduced that reported an error in case
someone changes default reprcode from the template in the object, but
doesn't change the value itself (leaving it to be default).
But it hasn't accounted for the case where reprcode wasn't actually
changed, but repeated to be the same as in template. We don't want any
error in this case.
Fixed and added test.