Make types a class attribute in dlis
dlis.types acts as a parsing guide for how object-types from the
.dlis-file should be parsed as python objects. To continue on the
journey where we allow for users to customize how dlisio parses the
file, types is made available to the end-user.

Effectively this gives the end-user the opportunity to add new vender-
specific (or other) object-types that will be parsed as if they where a
standard object-type from dlisio.