Rename dlis.object_sets -> dlis.indexedobjects
The intention of dlis.object_sets is to gather all objects of the same
type, regardless of their original object set. In rp66 object sets
refer to a set of object with the same type and attributes. However it
supports multiple sets of the same type.  Hence the name can be a bit
confusing for those familiar with the rp66 standard.