Read multidimensional fdata correctly
4.4.3 DIMENSION attribute of RP66V1 defines the way multidimensional
data is sequentially stored in FDATA.
With the first index changing first, data are presented in column-head
way. Python, however, assumes row-head way of data presentation.

Hence the data was read incorrectly. For dimension [3, 2] and linear
data 1, 2, 3, 4, 5, 6 the expected table is

1 4
2 5
3 6

So [1, 4], [2,5], [3, 6] or [1, 2, 3], [4, 5, 6] representations were
expected.
But the data were read into [[1, 2], [3, 4], [5, 6]].

Hence in python interface we reverse dimension, axis, and
element-limit.
Reversed dementions become [2, 3] (2 lists with 3 elements in each),
instead of declared [3, 2] (3 lists with 2 elements in each).

Thus data will be read as [[1, 2, 3], [4, 5, 6]].

To reach original values user has to use the attic property.