Preserve unrecongnized objects in refs
Looks like decision not to include unrecognized links to the actual link
attributes has not been the best one.
Order matters for certain link objects. For example, AXIS objects
correspond to related DIMENSION objects. If reference to some of the
AXIS objects is missing, but reference to DIMENSION object is there,
shift happens, and DIMENSION[i] doesn't correspond anymore to AXIS[i].
Hence None values will be inserted in place of not-found objects.