Change valuetype for measurement max/std deviation
Table in RP66 restrics Maximum Deviation and Standard Deviation in
Calibration-Measurement objects to count 1.
But according to corresponding comments for these attributes as long as
for Dimension attribute, this is not true.
Both Maximum-deviation and standard-deviation must contain only 1
sample, but that sample should correspond to the dimensions of the
Dimension attribute. So count should correspond to the values of the
Dimension attribute.

Note that in RP66 V2 this is fixed and restriction is lifted.
Hence assuming precedence of the comment and allowing more than 1 value
to be present there.

Updated test file.

However, situation with Standard attribute is not so clear. Dependency
on Dimension is mentioned in RP66 V1, but doesn't seem to make a lot of
sense from the context. As well, in RP66 V2 it doesn't appear in the
description of the Dimension attribute.
Hence considering this to be a mistake.