Remove initial value for dlis.types in docs
By default sphinx renders the value of class attributes. For dlis.types,
which is a dict of (create)methods, the output becomes very verbose.

There is no simple way of removing defaultvalues of class attributes,
hence this ugly hack where the attribute is excluded and then added
manually with a custom :annotation: