Allow invalid reprcodes in template
Change accounts for certain files that violate RP66.

In seen files, representation code in templates was invalid, in
particular - 0. Representation code was overwritten to correct value
only in actual objects.

Right now dlisio fails to read such files because we fail on noticing
invalid representation codes.
To allow processing of files like that, change is made to validate
representation code only when value is or should be present.