Interface changes for sampled attributes
 - Remove automatic zoning of value attributes [1]
 - Use skip() for attic entries that are accessed through properties
 - limit measurement attributes, std/max-deviation, reference and
   plus/minus-tolerance to 1 sample [3]

[1] After some considerations, indexing values by zone may not be the
best interface we can provide the user.

Although the standard defines that the values attribute from Parameter
and Computation should be interpreted by corresponding zones, it seems
that the most common usecase is that there is no zone and values is
scalar or a linear array. Hence indexing by zones more often than not a
gives little or none information, while the interface is more clunky to
use.

We should aim to accomodate the most common usecases in a sensible way,
rather than design for the complex cases that the standard support. The
user can still make the connection between values and zones manually for
the cases where it is desired.

[2] Use skip() on all attic entries that are accessed through properties
in the python object. skip() makes the parsing dict more intuitive as we
explicity state that we do not care about these attributes at this point
in time. Also, we avoid that these entries end up in stash.

[3] The standard is a bit unclear about how these attributes should be
interpreted, but we have concluded that they should only contain one
sample, where the sample dimension is defined by the dimension
attribute.