Make objects attribute a property
Change allows us to have only one access point to the data (through
indexedobjects dictionary).
It's still might be convenient for users to have direct access to all
objects, but these use cases are quite rare, hence it shouldn't be a big
computation overhead to recreate this dictionary on fly.