Use std::transform not raw loop
The main insentive for this change is to shut up cppcheck. In a general
sence cppcheck is right to flag this loop as a style warning. However in
this spesific case, where we operate on the full range of the vector
anyways, one can argue that the for loop is cleaner. BUT than a inline
suppression is needed, hence the switch to std::transform is prefered.