Allow building python core with setup.py
This essentially entails that the user can install the python core with
setup.py directly, assuming all dependancies are installed. In effect
mio and mpark need to be considered to be external requirements.

The normal build sequence makes use of the same configuration to build
the core. However, here mio and mpark is added as subdirectories to the
project and was not exported to the build tree.

To unify the to cases, the normal build sequence need to export mio and
mpark to the build tree when building the c++ core. Then set the
appropriate environment variables so cmake can find them when when
building the python core.