Make extension/typeinfo::name char[], not char*
The dl::typeinfo struct is used to look up a DLIS primitive's
representation code and string representation. The string representation
is constexpr and can be char[], meaning its size is still available.

This is necessary to use it as a description in signatures and
docstrings generated by pybind11 starting from pybind11 v2.3 [1]

[1] https://github.com/pybind/pybind11/issues/1927