Add fdata tests
Next fdata issues are planned to be fixed in the near future:
- Allow all the reprcodes in fdata
- Allow many fdatas in one fdata ILR
- Allow non-sequential frame numbers

Following decisions have been made by team regarding this commit:
- We prefer fast and simple merged test files over dynamic creation of
flexible and easily extendible ones, as tests won't ever change due to
fixed RP66 format.
- Tests are written experimental TDD style to suite the future
implementation as expected by the team.
- Test code duplication is not a bug, it's a feature to ensure that
reporting of every test failure is nice and easy-to-read. Also tests are
not expected to change.