Refactor and add docstrings to dlis class
The dlis class represents logical files. It is one of the main
interactions for the end-user, hence it should be properly documented.

This class exposes many attributes and functions that are mainly
intended for internal use (but are deliberately made available for
advanced use-cases and debugging). This made the current class docs a
bit confusing. Thus the order of it's attributes, properties and methods
are reordered for clarity.