Add index channel as property on Frame and Channel
When working with curves its integral to know how they are indexed.
From the standard, identifying the index channel for a Frame or another
channel is a bit clumsy:

> The first channel in a Frame is implicitly the index channel _if_
  Frame.index_type is not None. Otherwise, all channels in that frame is
  index by samplenumber.

Adding the index as a property to Frame and Channel relieve the end user
of these details and ensures it is handled correctly.