Add dlis_packflen function
When working with records with variable-length data types, it's
necessary to first figure out how many bytes both the source and
destination is when packing.

The packflen function is analoguous to strlen, and solves this exact
problem.

In general, it will always be necsesary to either over-allocate, or scan
the input twice. This can be avoided if it's known in advance that the
input or output is fixed-sized, but in general packflen should always
work.

Some refactoring was necessary, since figuring out the size is virtually
the same thing as reading and transforming.

The packf tests are augmented to check the packflen.