Compute fdata frame size from the frame
Instead of assuming that the format is fixed-size, figure out the size
for every frame as it is read. This is a performance hit for fixed-size
records, but enables reading data types of variable size, e.g. in the
presence of uvari and ascii.

Please not that only the computation of size is changed - the
no-variable-size restriction on reading fdata is not lifted, as it
depends on reallocation support.

This means that this commit does not introduce any observable change,
but the tests still pass, demonstrating that the approach works for
fixed-size records (and then likely variable-sized too).