Replace fixed-size check with supported-fmt list
Instead of implicitly checking for supported formats while also
computing the size of the pre/fmt/post parts of the format, explicitly
check for supported formatting options.

Supporting a format consists of both making space and declaring the type
in python (.py), in addition to shuffling the bytes in the right order
(ext/.cpp). An explicit supported-check allows this work to go
gradually, adding support for some variable-sized types without having
to support all of them at once.