Remove now-obsolete check_supported_fmtstr
The check_supported_fmtstr existed as a filter/feature flag for
failing unsupported types in FDATA frames, but now that all data types
are supported, this check is nonsensical.

There could still be use for a C++ check that the format string is
valid, but it is distinct from this function which raised
NotImplementedError.