Read attributes directly from attic
Drop the 2-phase load() and link() object creation. Initialize object
with attic and let properties parse values from attic on-demand. This
solves several issues with the 2-phase initialization process:

 - object-to-object references are often cyclic by nature, which
   complicated the dependency graph between the python objects. Object
   references are now handled by the attribute itself, which does a
   just-in-time look-up in the logical file's object-pool.

 - The state of the object is now clear. The 2-phase initialization
   required us to keep track of the current state of each object. I.e.
   has it been loaded with load()? Has references been created with
   link()?

 - There is no longer a requirement that _all_ objects needs to loaded,
   and then linked in dlisio.load(). By not creating the
   object-to-object references in dlisio.load() the execution time is
   reduced to as much as 35%, depending on the number of objects in the
   file.

 - Hot-swapping objects during the lifetime of the file-handle. Loading
   more objects into the object-pool post dlisio.load() would also need a
   reload of _all_ current object to look for new potential updated.
   With a just-in-time lookup into the object-pool, this is no longer an
   issue.

Additionally this change makes it possible for the dlisio.load() to
parse just a subset of the file's objects, yielding even greater
performance. The implementation of this is outside the scope of this
commit.

The use of properties means that it is no longer posible to set
attribute values (with out setting them directly in attic).
property-setters would solve this but as we do not support writing of
files so this should be fine.

It also make the niche of monkey-patching new attributes more difficult
as properties need to be added to the class itself.