Remove skip-keyword from attributes
The intension behind skip was to tell load() not to parse this
attribute. This was handly for cases where the attribute that needed
spesial care, that could not be handled by a general load-rutine.

load() has been replaced by properties that handle parsing and linking
on demand, hence the skip-keyword serves no purpose anymore.