Add user-defined string decoding table
Strings in Python are UTF-8, and strings in DLIS [1] are supposed to be a
subset of ASCII, which is also valid UTF-8. However, a lot of file are
written with encoded strings. Previously this was handled by only
looking for the degree symbol from latin1 and replacing it with UTF-8
degree, but that's not a good strategy.

Now, ASCII (i.e. valid UTF-8) is still assumed, but users can provide a
table of encodings to try if it should fail with dlisio.set_encodings.

This means that this patch BREAKS on files where the degree symbol
(0xB0) was the only non-ascii character. This is unfortunate, but
overall healthy, because users must now explicitly say that an encoding
is fine, rather than our hacky guess. Some strings are valid in multiple
encodings, and it's a user responsibility to make sure the correct one
is picked.

There is no way to guess what encodings are acceptable, as there is
nowhere in the DLIS file where encoding can (according to spec) be
specified.

[1] IDENT, ASCII, and UNITS