Return pyobj.release() instead of inc_ref()
Use pybind11::object::release() rather than inc_ref() when returning
pointers to objects created in C++ when the function returns a handle.

The release() method will release and return the underlying pointer,
without decreasing the ref count. inc_ref() was always a trick to make
sure that destructor of pybind11::object had a ref count of 2 before the
pybind11::object destructor ran (and reduced it to 1) on frame teardown,
since the handle does not concern itself with ref count.

Turns out this behaviour was always available through release(), which
has the performance benefit of not having to touch the ref count at all.