Integrate FRAMNO into FDATA/curves
Instead of treating frameno as some magic sometimes-fails number,
integrate it into the curves array. If the frameno is always available,
it's easy to detect missing frames, handle out-of-order or even
duplicate framenos in user code, rather than trying to get it right in
dlisio (and surely failing).

A huge set of tests needs updating, but it's largely just updating the
index. The first column in the curves matrix is now FRAMENO, not whever
interesting data that was there before.

This patch is quite likely to break a lot of code out there, but is
probably for the better in the long run.