Support multiple FDATA/rows in one IFLR
RP66 explicitly allows there to be more FDATA/rows in a single IFLR.
Support for it was pushed, because it requires resizing of arrays, as
there is no good way to know in advance how many rows there are, when
they don't correspond to the IFLRs.

It's still read_fdata that powers this, through resizing of the numpy
array when there are more rows in a single record. It's a somewhat
intricate process to do this resizing, rather than using something
simpler like std::vector, but it's done in order to maintain exception
safety, so that objects don't leak should parsing a row fail.