Parse objectsets on demand
Usually files contains a lot of metadata that the user never tries to
access or have interest in. Hence loading objects on a per-type basic
makes more sense than to load everything when opening the file.

By default only FILE-HEADER, ORIGIN, FRAME and CHANNEL objects are
loaded, while other object-types are loaded when needed.  This change is
not observable by the end-user as the dlis class takes care of loading
objects whenever they are needed.

Parsing objectsets is quite expensive so this change introduce a huge
performance gain for opening files with a lot of metadata objects.

Together with [1] this patch reduces the time it takes to load a 31M
file with ~30k objects from ~2.7sec to ~0.7sec on my computer.

[1] SHA: 3647bb008eb1e4a38d9281abf16f2f75832c0d08