Return mnemonic not Channel for frame.index
With FRAMENO integrated into the numpy.ndarray returned by .curves()
it's natural to return FRAMENO, not None, from .index if there is no
index Channel. However, FRAMENO is not a Channel object. To make the
index properties coherent with FRAMENO, Frame.index now returns the
channel mnemonic, rather than the full channel object.

This is a breaking change, and some workflows where the user works
directly with Channels and not through the Frame objects will be a bit
more cumbersome. However, index now works seamlessly with indexing the
.curves() array, e.g:

>>> curves = frame.curves()
>>> curves[[frame.index, 'GR', 'ALH1']]