Remove Channel.index
Channel.index was introduced as a convenience for users working directly
with a 'stand-alone' Channel. It was an attempt at abstracting away the
need to interact directly with the Channels Frame. In hindsight this was
a poor decision as the Frame supply more context to the Channel than
just the index. Hence a Channel such not be considered as a stand-alone
piece of information.

Removing Channel.index forces the users to interact with the full Frame,
as they should.