Warn when file contains UPDATE-object(s)
The lack of support for UPDATE-objects may have huge implications for
how information in the logical file is interpreted both by dlisio
internally and by the user.

It will require a significant effort to robustly support UPDATE-objects
in dlisio. However, it is unclear how frequent these objects are
present, so for now just warn if UPDATE objects are present in the
logical file.