Fix a bug where match() yielded the same obj more than once
The dlis class keeps track of the type of object_sets that are present
in the file. This list is being used by dlis.match to match on
object-type. Because this is a list, and not a set, match would search a
particular object-type multiple times if there where multiple
object_sets of that type.