Reintroduce object-counts in dlis.describe()
Due to the just-in-time loading of objects introduced in [1], the
object-count for each object-type in dlis.describe() was removed as it
would in effect load every object in the file.

Seaming that this information is valuable it is re-introduced on the
grounds that its fair to assume that describe() is only called in the
context of exploring a single file, or a few files. Hence, we can
tolerate the performance hit.

[1] SHA: f5090cc67b1d12add8ca4dab44bd0278bcde01f9