Explicity unmap file in dlisio.load
If dlisio.load should fail for some reason, it left the memory mapping
of the file open [1].

dlisio.load relies on the fact that the destructor of the underlying
mio::mmap object unmapped the file. This teardown is left to Python's
GC, which apperently didnt always do a great job.

To be sure to avoid dangling filehandles, we explictly unmap the file in
dlisio.load().

[1] https://github.com/equinor/dlisio/issues/186