Allow duplicated mnemonics in curves()
By default we allow duplicated mnemonics as long as Channels are
distinguishable by the triplet; mnemonic, origin, copynumber. If that is
not the case, its a clear standard violation and curves() throws.

Apparently there are files where this is the case, hence we supply the
option to read the curve-data anyway. The default-behavior of curves()
is still to throw if the standard is violated.