Force channel.curves() to return copy when slicing
Because all channel.curves() does is to call frame.curves() and slice
the result, we should return a copy of the slices array rather than a
view into the original array. The reason being that a view keeps the
original array alive as long as the view live, meaning we lay claim to
way more memory than needed.

Copying is slower than returning a view, but it can potentially save so
much memory that the copy is worth it. Further down the line, the core
library will support reading sub-chunks of frames, which will make the
extra copy obsolete.